REST API Client was syncronized with last changes of the specification
