+++ package.json
@@ -7,6 +7,7 @@
     "scripts": {
         "build": "rollup -c",
         "release": "cross-env NODE_ENV=production rollup -c",
+        "generate-api-client": "npx swagger-typescript-api -p ./src/api/client/swagger.json -o ./src/api/client -n client.ts",
         "start": "rollup -c -w",
         "test": "jest"
     },
