+++ src/api/client/index.ts
@@ -1,12 +1,11 @@
-import { Api, Feedback, HttpResponse } from './api';
+import { Api, HttpResponse } from './client';
 
 const client = new Api();
 
 export function postFeedback(video_id: string, distractful: boolean) : Promise<HttpResponse<void, any>> {
-    let fb: Feedback = { video_id, distractful };
-    return client.api.feedback(fb)
+    return client.api.feedbackCreate({ video_id, distractful })
 }
 
-export function filterVideos(videoIds: string[]) : Promise<HttpResponse<string[], string>> {
-    return client.api.filterVideos(videoIds);
+export function filterVideos(videoIds: string[]) : Promise<HttpResponse<{distractful_ids: string[] }, string>> {
+    return client.api.filterDistractfulVideosCreate({ recommended_ids: videoIds });
 }
